+++ src/pages/user/MyPage.tsx
@@ -1,0 +1,5 @@
+const MyPage = () => {
+  return <div>MyPage</div>;
+};
+
+export default MyPage;
