+++ src/pages/user/UserPage.tsx
@@ -1,0 +1,5 @@
+const UserPage = () => {
+  return <div>UserPage</div>;
+};
+
+export default UserPage;
